修改方法返回类型，保证标签Id与其他地方一致

+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/api/WxMpUserTagService.java
@@ -94,8 +94,8 @@ public interface WxMpUserTagService {
    * 详情请见：<a href="http://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421140837&token=&lang=zh_CN">用户标签管理</a>
    * 接口url格式： https://api.weixin.qq.com/cgi-bin/tags/getidlist?access_token=ACCESS_TOKEN
    * </pre>
-   *
+   * @return 标签Id的列表 
    */
-  List<Integer> userTagList(String openid) throws WxErrorException;
+  List<Long> userTagList(String openid) throws WxErrorException;
 
 }
